requirements.txt
rich
GitPython
lizard
dockerfile-parse
transformers
